move slash command handling into chatwidget
